requirements.txt
Brotli==1.0.9
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
ConfigArgParse==1.5.3
dnspython==2.2.1
Faker==13.14.0
faker-enum==0.0.2
Flask==2.1.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==1.5.5
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
locust==2.10.0
MarkupSafe==2.1.1
msgpack==1.0.4
numpy==1.23.0
psutil==5.9.1
pymongo==4.1.1
python-dateutil==2.8.2
python-snappy==0.6.1
pyzmq==22.3.0
requests==2.28.1
roundrobin==0.0.2
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
